requirements.txt
# Create requirements.txt
cat > requirements.txt << EOF
# Core dependencies
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
opencv-python==4.8.1.78

# Vector database and embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# Transformers and ML
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
accelerate==0.25.0

# API framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6

# Frontend
streamlit==1.29.0

# Utilities
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1

# Development tools
black==23.12.1
flake8==6.1.0
ipython==8.18.1
EOF

# Install all dependencies
pip install -r requirements.txt